requirements.txt
django-cassandra-engine==1.9.0
Django==3.2.4